+++ src/Command/CommandInterface.php
@@ -11,6 +11,14 @@ interface CommandInterface
     public function getCommand();
 
     /**
+     * Tells if this command blocks while waiting for a response, to avoid
+     * being affected by connection timeouts.
+     *
+     * @return bool If true, this command blocks
+     */
+    public function isBlocking();
+
+    /**
      * Get processed arguments for command
      *
      * @return array Arguments