+++ tests/Command/AckJobTest.php
@@ -22,6 +22,13 @@ class AckJobTest extends PHPUnit_Framework_TestCase
         $this->assertSame('ACKJOB', $result);
     }
 
+    public function testIsBlocking()
+    {
+        $c = new AckJob();
+        $result = $c->isBlocking();
+        $this->assertFalse($result);
+    }
+
     public function testBuildInvalidArgumentsEmpty()
     {
         $this->setExpectedException(InvalidCommandArgumentException::class, 'Invalid command arguments. Arguments for command Disque\\Command\\AckJob: []');