+++ tests/Command/DequeueTest.php
@@ -22,6 +22,13 @@ class DequeueTest extends PHPUnit_Framework_TestCase
         $this->assertSame('DEQUEUE', $result);
     }
 
+    public function testIsBlocking()
+    {
+        $c = new Dequeue();
+        $result = $c->isBlocking();
+        $this->assertFalse($result);
+    }
+
     public function testBuildInvalidArgumentsEmpty()
     {
         $this->setExpectedException(InvalidCommandArgumentException::class, 'Invalid command arguments. Arguments for command Disque\\Command\\Dequeue: []');